change: upgrade ability system

+++ Assets/Scripts/Signals/TookDamageSignal.cs
@@ -5,6 +5,6 @@ namespace Signals
     public class TookDamageSignal
     {
         public IEntity Entity { get; init; }
-        public long    Damage { get; init; }
+        public float   Damage { get; init; }
     }
 }
